backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.33.0
sqlalchemy==2.0.43
pymysql==1.1.1
python-jose[cryptography]==3.4.0
python-dotenv==1.0.1
pydantic-settings==2.8.1
email-validator==2.2.0
numpy==1.24.4
Pillow==10.4.0
torch==2.1.2
transformers==4.38.2
openai==1.51.2
requests==2.32.3
